big refactor
back to green
