+++ src/WP_CLI/AssocArgs.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WP_CLI_Dotenv_Command;
+namespace WP_CLI_Dotenv\WP_CLI;
 
 class AssocArgs
 {
