+++ src/functions.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace WP_CLI_Dotenv_Command;
-
-use WP_CLI;
+use WP_CLI_Dotenv\Dotenv\File;
 
 /**
  * @param $key
  * @param $value
  *
  * @return string
+ * 
+ * @deprecated 
  */
 function format_line($key, $value)
 {
     if (strpos($value, ' ') !== false) {
         $value = "'$value'";
     }
-    return sprintf(Dotenv_File::LINE_FORMAT, $key, $value);
+    return sprintf(File::LINE_FORMAT, $key, $value);
 }
 
 /**
@@ -24,6 +24,8 @@ function format_line($key, $value)
  * @param $file
  *
  * @return string
+ * 
+ * @deprecated
  */
 function get_filepath($file)
 {
@@ -48,14 +50,16 @@ function get_filepath($file)
  *
  * @param $args
  *
- * @return Dotenv_File
+ * @return File
+ * 
+ * @deprecated
  */
 function get_dotenv_for_read_or_fail($args)
 {
     $filepath = get_filepath($args);
 
     try {
-        $dotenv = Dotenv_File::at($filepath);
+        $dotenv = File::at($filepath);
         return $dotenv->load();
     } catch (\Exception $e) {
         WP_CLI::error($e->getMessage());
@@ -68,14 +72,14 @@ function get_dotenv_for_read_or_fail($args)
  *
  * @param $args
  *
- * @return Dotenv_File
+ * @return File
  */
 function get_dotenv_for_write_or_fail($args)
 {
     $filepath = get_filepath($args);
 
     try {
-        $dotenv = Dotenv_File::writable($filepath);
+        $dotenv = File::writable($filepath);
         $dotenv->load();
     } catch (\Exception $e) {
         WP_CLI::error($e->getMessage());
@@ -95,10 +99,8 @@ function get_dotenv_for_write_or_fail($args)
 function prompt($question, $default)
 {
     try {
-        $response = \cli\prompt($question, $default);
+        return \cli\prompt($question, $default);
     } catch (\Exception $e) {
         WP_CLI::error($e->getMessage());
     }
-
-    return $response;
 }
