+++ tests/unit/Dotenv_LineTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use WP_CLI_Dotenv_Command\Dotenv_Line;
+use WP_CLI_Dotenv\Dotenv\Line;
 
 class Dotenv_LineTest extends PHPUnit_Framework_TestCase
 {
@@ -9,7 +9,7 @@ class Dotenv_LineTest extends PHPUnit_Framework_TestCase
      */
     function it_has_a_method_for_getting_the_key()
     {
-        $line = new Dotenv_Line('FOO=BAR');
+        $line = new Line('FOO=BAR');
         $this->assertSame('FOO', $line->key());
     }
 
@@ -18,7 +18,7 @@ class Dotenv_LineTest extends PHPUnit_Framework_TestCase
      */
     function it_has_a_method_for_getting_the_value()
     {
-        $line = new Dotenv_Line('FOO=BAR');
+        $line = new Line('FOO=BAR');
         $this->assertSame('BAR', $line->value());
     }
 
@@ -27,7 +27,7 @@ class Dotenv_LineTest extends PHPUnit_Framework_TestCase
      */
     function it_can_return_itself_as_a_string()
     {
-        $line = new Dotenv_Line('FOO = Monkey stuff');
+        $line = new Line('FOO = Monkey stuff');
         $this->assertSame('FOO=Monkey stuff', $line->toString());
     }
 
@@ -36,9 +36,9 @@ class Dotenv_LineTest extends PHPUnit_Framework_TestCase
      */
     public function it_strips_matching_quotes_from_both_ends_of_the_value()
     {
-        $lineWithSingleQuotedValue = new Dotenv_Line("FOO = 'BAR'");
-        $lineWithDoubleQuotedValue = new Dotenv_Line('FOO = "BAR"');
-        $lineWithMixedQuotedValue = new Dotenv_Line('FOO = \'BAR"');
+        $lineWithSingleQuotedValue = new Line("FOO = 'BAR'");
+        $lineWithDoubleQuotedValue = new Line('FOO = "BAR"');
+        $lineWithMixedQuotedValue = new Line('FOO = \'BAR"');
 
         // value is wrapped with double quotes
     	$this->assertSame('BAR', $lineWithSingleQuotedValue->value());
@@ -47,4 +47,16 @@ class Dotenv_LineTest extends PHPUnit_Framework_TestCase
         // value is wrapped with quotes that do not match - could be part of value itself
         $this->assertSame('\'BAR"', $lineWithMixedQuotedValue->value());
     }
+
+    /**
+     * @test
+     */
+    public function it_can_be_constructed_from_a_pair()
+    {
+    	$line = Line::fromPair('key', 'value');
+
+        $this->assertSame('key', $line->key());
+        $this->assertSame('value', $line->value());
+    }
+
 }
