+++ tests/unit/Dotenv_LinesTest.php
@@ -1,18 +1,18 @@
 <?php
 
+use WP_CLI_Dotenv\Dotenv\FileLines;
 use Illuminate\Support\Collection;
-use WP_CLI_Dotenv_Command\Dotenv_Lines;
 
-class Dotenv_LinesTest extends PHPUnit_Framework_TestCase
+class FileLinesTest extends PHPUnit_Framework_TestCase
 {
-    use Fixtures;
+    use WP_CLI_Dotenv\Fixtures;
 
     /**
      * @test
      */
     function it_is_a_collection()
     {
-        $this->assertInstanceOf(Collection::class, new Dotenv_Lines);
+        $this->assertInstanceOf(Collection::class, new FileLines);
     }
 
     /**
@@ -21,7 +21,7 @@ class Dotenv_LinesTest extends PHPUnit_Framework_TestCase
     function it_has_a_method_for_loading_from_a_file()
     {
         $filepath = $this->get_fixture_path('env-basic');
-        $lines = Dotenv_Lines::fromFile($filepath);
+        $lines = FileLines::load($filepath);
 
         $this->assertEquals(count(file($filepath)), $lines->count());
     }
@@ -32,7 +32,7 @@ class Dotenv_LinesTest extends PHPUnit_Framework_TestCase
     function it_can_return_itself_as_a_single_string()
     {
         $filepath = $this->get_fixture_path('env-basic');
-        $lines = Dotenv_Lines::fromFile($filepath);
+        $lines = FileLines::load($filepath);
         $contents = file_get_contents($filepath);
 
         $this->assertSame($contents, $lines->toString());
@@ -44,7 +44,7 @@ class Dotenv_LinesTest extends PHPUnit_Framework_TestCase
     function it_can_return_itself_as_a_collection_of_key_value_pairs()
     {
         $filepath = $this->get_fixture_path('env-basic');
-        $lines = Dotenv_Lines::fromFile($filepath);
+        $lines = FileLines::load($filepath);
 
         $this->assertSame('FOO', $lines->pairs()->first()->key());
         $this->assertSame('BAR', $lines->pairs()->first()->value());
@@ -57,7 +57,7 @@ class Dotenv_LinesTest extends PHPUnit_Framework_TestCase
     // function it_parses_each_line_into_an_array_of_its_parts()
     // {
     //     $filepath = $this->get_fixture_path('env-basic');
-    //     $lines = Dotenv_Lines::fromFile($filepath);
+    //     $lines = FileLines::fromFile($filepath);
     //
     //     $keys = ['key','value','type'];
     //     $this->assertArraySubset($keys, array_keys($lines[0]));
